refactor: 将app-theme-provider的工具函数迁移至utils\theme.ts
